Reorganizing test files and initial project setup
