+++ .gitignore
@@ -1,0 +1,10 @@
+## Byte-compiled, execution cache
+__pycache__/
+tests/dask-worker-space
+examples/dask-worker-space
+
+# Distribution / packaging
+.eggs/
+eggs/
+*.egg-info/
+
