+++ hipscat/catalog.py
