+++ hipscat/util.py
