+++ setup.bash
@@ -1,0 +1,2 @@
+
+export PYTHONPATH=$PWD/hipscat:$PYTHONPATH
